+++ src/agat/agat_sp_add_introns/test.sh
@@ -8,7 +8,7 @@ set -eo pipefail
 test_dir="${meta_resources_dir}/test_data"
 
 # create temporary directory and clean up on exit
-TMPDIR=$(mktemp -d "$meta_temp_dir/$meta_functionality_name-XXXXXX")
+TMPDIR=$(mktemp -d "$meta_temp_dir/$meta_name-XXXXXX")
 function clean_up {
   [[ -d "$TMPDIR" ]] && rm -rf "$TMPDIR"
 }