+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 * Replace the deprecated use of the meta variable `functionality_name` by just `name` (PR #174).
 
+* Bump viash to `0.9.4` (PR #).
+
 # biobox 0.3.0
 
 ## NEW FUNCTIONALITY