+++ CHANGELOG.md
@@ -6,6 +6,12 @@
 * `cellranger/cellranger_count`: Align fastq files using Cell Ranger count (PR #163).
 
 
+## MINOR CHANGES
+
+* Replace the deprecated use of the meta variable `functionality_name` by just `name` (PR #174).
+
+* Bump viash to `0.9.4` (PR #175).
+
 # biobox 0.3.0
 
 ## NEW FUNCTIONALITY